SALTO-7283: Protect Static file from long file name errors
