+++ packages/adapter-api/package.json
@@ -32,7 +32,8 @@
     "@salto-io/dag": "0.5.0",
     "@salto-io/logging": "0.5.0",
     "@salto-io/lowerdash": "0.5.0",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "truncate-utf8-bytes": "^1.0.2"
   },
   "devDependencies": {
     "@jest/expect-utils": "^29.7.0",
