+++ packages/adapter-utils/package.json
@@ -40,7 +40,6 @@
     "joi": "^17.4.0",
     "lodash": "^4.17.21",
     "object-hash": "^3.0.0",
-    "truncate-utf8-bytes": "^1.0.2",
     "wu": "^2.1.0"
   },
   "devDependencies": {
