+++ packages/adapter-utils/src/nacl_case_utils.ts
@@ -5,19 +5,11 @@
  *
  * CERTAIN THIRD PARTY SOFTWARE MAY BE CONTAINED IN PORTIONS OF THE SOFTWARE. See NOTICE FILE AT https://github.com/salto-io/salto/blob/main/NOTICES
  */
-import path from 'path'
-import truncate from 'truncate-utf8-bytes'
 import invert from 'lodash/invert'
-import { hash as hashUtils } from '@salto-io/lowerdash'
+import { MAX_PATH_LENGTH } from '@salto-io/adapter-api'
 
 const NACL_ESCAPING_SUFFIX_SEPARATOR = '@'
 const NACL_CUSTOM_MAPPING_PREFIX = '_'
-// Windows has the lowest known limit, of 255
-// This can have an effect at a time we add a ~15 chars suffix
-// So we are taking an extra buffer and limit it to 200
-const MAX_PATH_LENGTH = 200
-const MAX_PATH_EXTENSION_LENGTH = 20
-
 const allCapsRegex = /^[A-Z]+$/
 const camelCaseRegex = /[a-z][A-Z]/g
 const allCapsCamelCaseRegex = /[A-Z]([A-Z][a-z])/g
@@ -25,22 +17,6 @@ const allCapsCamelCaseRegex = /[A-Z]([A-Z][a-z])/g
 export const pathNaclCase = (name?: string): string =>
   (name ? name.split(NACL_ESCAPING_SUFFIX_SEPARATOR)[0] : '').slice(0, MAX_PATH_LENGTH)
 
-// Trim part of a file name to comply with filesystem restrictions
-// This assumes the filesystem does not allow path parts to be over
-// MAX_PATH_LENGTH long in byte length
-export const normalizeFilePathPart = (name: string): string => {
-  if (Buffer.byteLength(name) <= MAX_PATH_LENGTH) {
-    return name
-  }
-  const nameHash = hashUtils.toMD5(name)
-  let extension = path.extname(name)
-  if (extension.length > MAX_PATH_EXTENSION_LENGTH || Buffer.byteLength(extension) !== extension.length) {
-    // Heuristic guess - a valid extension must be short and ascii
-    extension = ''
-  }
-  const suffix = `_${nameHash}${extension}`
-  return truncate(name, MAX_PATH_LENGTH - suffix.length).concat(suffix)
-}
 // Current values in this mapping should not be changed
 // Values in the map should be unique
 // Adding more values should be with a leading z as an indication the value has more than one letter
